Move workers to new dir
