+++ app/lib/workers/Worker.svelte.ts
