+++ app/routes/dashboard/library/+page.svelte
@@ -17,11 +17,11 @@
 	import { useSidebar } from '$lib/components/ui/sidebar'
 	import { LibrarySymbolGroups, LibrarySymbols, LibrarySymbolFields, LibrarySymbolEntries, manager, SiteSymbols } from '$lib/pocketbase/collections'
 	import type { LibrarySymbol } from '$lib/common/models/LibrarySymbol'
-	import { useImportLibrarySymbol } from '$lib/ImportSymbol.svelte'
+	import { useImportLibrarySymbol } from '$lib/workers/ImportSymbol.svelte'
 	import { tick } from 'svelte'
 	import { BlockEditor } from '$lib/builder/views/modal'
 	import type { ObjectOf } from '$lib/pocketbase/CollectionMapping.svelte'
-	import { useExportLibrarySymbol } from '$lib/ExportSymbol.svelte'
+	import { useExportLibrarySymbol } from '$lib/workers/ExportSymbol.svelte'
 
 	const active_symbol_group_id = $derived(page.url.searchParams.get('group'))
 	const active_symbol_group = $derived(active_symbol_group_id ? LibrarySymbolGroups.one(active_symbol_group_id) : undefined)
